Snap: Remove "help screen on first start" feature
Revert "SnapdragonCamera: Add help menu to camera2"

This reverts commit 0dbfdffc9994a17e60c57b17c3cc5743cc75e155.


Revert "SnapdragonCamera: Move help menu string to ressource"

This reverts commit 901130273dbfc753ef883941a4e93b50056177e4.


Revert "SnapdragonCamera: Adding help screen on first start"

This reverts commit 666692bf730a1c46210cc74781a56afb4b6fcf28.

+++ src/com/android/camera/ui/OneUICameraControls.java
@@ -309,7 +309,8 @@ public class OneUICameraControls extends RotatableLayout {
         } else {
             v.setY(mHeight - mBottom + (mBottom - h) / 2);
         }
-        float bW = mWidth / 5f;
+        int bW = mWidth / 5;
+
         v.setX(bW * idx + (bW - w) / 2);
     }
